Make sure the status bar can be used to correctly display the process
This makes significant changes to the way the visibility of phases are handled.

We confirm that users can "view" the phase. If not we then remove the phase
from the list of phases they are shown. We then make sure that we always
have a valid list of phases when we build the status bar