Pull actions sidebar block into main template
We want applicants to have some actions on a Project so we need to have
the actions block in the main template for them to use.  However still
want the vast majority of action buttons & forms to be admin only so
this adds a few extra blocks to allow us the ability define those in a
more granular fashion.