Allow passing a path to the slack notification system
User.get_absolute_url was pointing at the wagtail admin which would expose the
admin urls to users on the frontend. The admin view is not necessarily the
"canonical" view for the user so should not be relied upon in the long run.

By allowing passing of a url to the slack notifier we allow the emitter to point
to the destination regardless of the more front end focused model method.