Feature/post demo 2 fixes (#1501)
* Make the choose file link intelligent

* Update the document chooser to user the model form and update tests

* Update the messaging to be more in line with other messages

* Make sure we tick the documents when all uploaded

* Tidy up the visual components (contract/links) on the detail page

* Tidy up how forms are contributed to the parent class

* Improve the delegated form base behaviour

* Refactor to remove the payments mixin

* Give the detail and edit views of the payment request their own view

* Remove the download link and change to a view

* Refactor the views to be better sorted

* Improve the interaction experience for payment requests and add activity

* Improve look of payment table to be a bit more bulked out

* Allow passing a path to the slack notification system

User.get_absolute_url was pointing at the wagtail admin which would expose the
admin urls to users on the frontend. The admin view is not necessarily the
"canonical" view for the user so should not be relied upon in the long run.

By allowing passing of a url to the slack notifier we allow the emitter to point
to the destination regardless of the more front end focused model method.