Remove Our Work section from Landing Page
Remove the Our Work section from the public Landing Page. The client did
not want this section. Instead of removing it outright, this patch
simply comments out the code that generates that section. Commenting
instead of removing will allow us to add the section back if the client
changes their mind in the future.