Hide the admin button for non-admins
There are several types of user with access to the review dashboard, but
only administrators are allowed to view the administrator pane.

This updates the admin dashboard to only render the admin button for
users with admin permission.  It also adds an ID to the button so
that test can reliably check for the presence of the button.

Issue #2724