Fix function call `AnonymousUser` performed in argument defaults (#3095)
It is generally not a good idea to perform function calls in argument
defaults in Python because the function will be called at the time the
default value is defined, not when the function is called. This means
that if the function has side effects, they will be executed every time
the default value is used, which can lead to unexpected behavior.