Fix search issue on submission list page with none archived value (#3125)
Fixed sentry issue
[3887915131](https://sentry.io/organizations/otf/issues/3887915131/?project=1268368)

#3123 has fixed it only for the users who can't see the archived feature
and that's why we found it working in our testing but for staff admin,
this issue is still there.

The main issue is we have checked for empty and None archived values but
with the search only it passes 'None' as the str type. I fix it by not
passing the archived param to the URL if the archived param's value is
None(not selected yet).