Use system fonts for the public and apply site
NOTE: The font files and `npm -> build:fonts` scripts are still kept as 
they are used by reportlab for rendering PDFs. 

If not using custom fonts, the pdfs are rendered with
Helvetica. Let’s discuss whether we need to use custom fonts for the PDFs.

The system font is not removed from the public site to keep up with the OTF branding. 

Closes #3131