Fixes regarding paf notifications  (#3102)
Partial fix of #3066
(https://github.com/HyphaApp/hypha/issues/3066#issuecomment-1337463266)

- [x] We should rename "/apply/projects/4/simplified/" to something like
"/apply/projects/4/approval/". Also rename
"ProjectDetailSimplifiedView".
- [x] The link to the uploads contract in the "Contracts" box should
open in a new tab, "target _blank"
- [x] When an applicant uploads a signed contract an e-mail is sent to
the applicant but not to the "Compliance Email" address from project
settings. I do not think we need the mail to the applicant but need to
add one to compliance.
- [x] When a user has approved a PAF I think the PAF should be removed
from that users dashboard.(Multiple tabs to show already approved and
approval required).
- [x] E-mail should be sent out to 'contracting+approver' user's email
address when all "PAF Reviewers Roles" have approved a PAF. The project
should also show in a "Final approving requeried" table on the Dashboard
for Contracting+Approver users.
- [x] Fix misspelling of "venor", should be "vendor"
- [x] Fix Permissions for contracts(hide buttons).
- [x] Do we need to separate "Compliance Email" setting to one address
for finance users and one for contracting users? Or should we send
directly to address is user accounts that has the role in question? (We
decided email notifications could go to a 'group email' and default to
individual emails if group email is not provided)

---------

Co-authored-by: sandeep <sandeep@Sandeeps-MacBook-Air.local>
Co-authored-by: Saurabh Kumar <theskumar@users.noreply.github.com>
Co-authored-by: Fredrik Jonsson <frjo@xdeb.org>