Use system fonts for apply site (#3132)
NOTE: The font files and `npm -> build:fonts` scripts are still kept as 
they are used by reportlab for rendering PDFs. 

The system font is not removed from the public site to keep up with the
OTF branding.

If not using custom fonts, the pdfs are rendered with
Helvetica. Let’s discuss whether we need to use custom fonts for the
PDFs.

Closes #3131