Update tests now that determinations can be reverted
Three tests removed due to those tests essentially testing that when a
submission is not in a determined state, and there's a determination,
the submission should end up in a determined state.  That's not true
anymore, as a submission can have a determination without being in a
determined state.

Added one test for the new workflow, allowing a revert.