update register flow to redirect back to application form
This PR ensure if the registeration is required before applying to a
fund then after registeration or after using forgot password the user
lands to the form submisison pages are password reset and/or register is
successful.

This PR also updates forgot password email template to be consistent with
other email format.

The flow to set the password after clicking on the activation link is
kept as it is.