Update register flow to redirect back to application form (#3521)
This pull request ensures that users are directed to the form submission
page after successfully registering or resetting their password if
registration is required before applying to a fund. Additionally, the
forgot password email template is updated to be consistent with other
email formats.

The flow to set the password after clicking on the activation link
remains unchanged.