Update typography to be consistent and scaled properly (#3478)
Use reference scale from tailwind to update the styles of different
headers.

Overall the fonts for headers were little too large and they are now
little scaled down so they don't look blown up.

Sensible formatting is applied to form elements, esp. help-text is made
smaller.

Adminbar component used on even more pages.

PS: A lot of files are touched and also a lot of code cleanup can still
be done. So for now the best way to test this would be to navigate
through the pages pages and find any obvious issues.