Fix wagtail user view pagination (#3613)
Closes #3607.

A quick one, just needed to have the variable changed from `page_obj` to
`users`, which is created and included in the context
[here](https://github.com/HyphaApp/hypha/blob/main/hypha/apply/users/admin_views.py#L167).
See screenshots below for a working example. Items per page was changed
from 20 to 5 to allow for demonstration. All other functionality works
as expected (ie. changing pages)