Update status bar to show statuses without hover (#3586)
Status Bar with latest designs.

The above status bars are from different submissions so don't get
confused with their status steps. The only change is that the statuses
are visible in the latest one while in the old one user has to hover
over every dot to know its status.