Added migration to remove NEW_SUBMISSION events without a submission (#3658)
Closes #3442. 

This small migration addition compliments #3445 as it will remove any
previous `NEW_SUBMISSION` events that belong to submissions that no
longer exist. This will allow for the removal of applicants that don't
have any active submission, but can't be deleted due to the `The object
you are trying to delete is used somewhere...` error.

Cheers!