Project Report Forms in Wagtail Admin
With `PROJECTS_ENABLED = True`, Project Report Forms may now be
created in Wagtail Admin and associated with a Fund. As of this commit
there is not yet the ability for applicants/grantees to respond to the
Form as entered.

The goal is to also use `PROJECTS_AUTO_CREATE = True` and possibly new
other settings to leverage Hypha Projects such that grantees will get
reminders about submitting Project Reports and have a place in Hypha
to send them.