Remove dependency of invoice delete on PROJECT_ENABLED Flag
If the PROJECT_ENABLED=False then the invoice should not be created at
the first place. But somehow at some point projects where enabled and
there are project and invoices in the system. The system should function
as expected compared to lacking functionality in/between.

This PR remove the PROJECT_ENABLED check on the project's
get_absolute_url, which is used while deleting invoices or potentially
at the other places.

closes #3663