Add "PAFs for review" table for paf approvers (#3645)
Fixes #3535 

Right now, we have two tables on the approvers dashboard "PAF waiting
for approval" and "PAF waiting for assignee" but as per the latest
design we want a new table "PAFs for review" that contains data from
both of the above tables and list it in a simple manner.
In this PR we are just adding that new table but not removing the older
ones, once we are done with the new table testing and are ready to have
it then we can remove the older ones.

The date requested column is kind of not clear yet. 
We have multiple dates, one of them is the "Submitted date" which is
when a lead/staff submits the PAF for the first time. Another one is
"Updated date" which is when staff submits or this approval(not paf) is
updated even for change approver/assignee(right now I have used this
one).
But I think we might need to maintain one more date as the "date
requested" for every approval, which might be different from both of the
above dates in sequential workflow(ex: the second approval
date_requested will be set as the first approval approved_date) while it
will be same as submit_date for parallel workflow.

What do you all think, is the submit_date enough, or should we have
date_requested as another as explained above?

**Update**
We decided to use submit_date as date_requested.
Removing the old tables as the new table contains both old tables' data.