Afford `PROJECTS_INITIAL_STATUS` to skip stages
When `PROJECTS_ENABLED` is true and the people running the Hypha
instance do not wish to use some of the first steps/stages of the
Projects capability, they can set `PROJECTS_INITIAL_STATUS` to a value
found in `apply/projects/models/project.py` `PROJECT_STATUS_CHOICES`.

This commit allows skipping straight to `invoicing_and_reporting`,
which is the only value that was confirmed to work alongside values of
`PROJECTS_ENABLED = True` and `PROJECTS_AUTO_CREATE = True`.