Enabled validation of determination draft saving (#3707)
Fixes #3706

A draft determination shouldn't be able to be saved without setting an
outcome, as there isn't a valid state to store it in (and if a
determination is being drafted, an outcome has been reached). This
enforces partial form validation when a determination draft is saved to
ensure an outcome is set.