Add server side previews to the application workflow (#3725)
## Description

Closes #3545.

This creates the option to require a user to preview their form before
submitting it. The requirement is set on a per fund/lab basis, and is a
~required checkbox when the lab/fund is configured in Wagtail admin~
global Django configuration option. The preview first saves the
application as a draft, then presents the user with a preview of their
application. Upon viewing, the user can either edit or submit as seen in
the flow diagram below.