Improve rtl support with modern css (#3732)
Fixes #3731

Improves rtl support among other things. Replaces attributes like
top/left/margin-right/padding-left as well as tailwind classes with
`*-inline*` and `*-block*` version.

`padding-inline-start` is for ltr the same as `padding-left` but for rtl
it is the same as `padding-right`. The browser just does the right
thing.

For tailwind we should always use `ms-1` instead of `ml-1` and `me-1`
instead of `mr-1` etc. The s/e stands for start/end.