Add Wagtail setting to skip vendor contract upload
Hypha tracks whether a contract has been signed and countersigned
depending on which role uploaded it and the context of the upload.
This change creates a new Wagtail setting called
`upload_countersigned_contract`. When set to `True`, whenever a
contract upload occurs, the UI shows that the document needs to be
countersigned and the document gets marked as countersigned. The net
effect is that there is no need for two contract uploads, only one.
When combined with the `staff_upload_contract` set to `True`, staff
can upload a single contract document without other users' actions.

Issue #3954