Display "View Live" button only if round is live in admin
Currently the "view live" button on the round detail admin page, links
to the round pages with the url containing fund slug followed by slug of
the round page.

The round can be accessed only via the fund url and the fund url serves
the active round, so the current link leads to 404.

This fixes the admin "view live" button so that it's displayed only if
the round is live and also links it correctly to the fund slug

Fixes #3794