Reduce the lazy load time for applicant submission on dashboard (#3815)
Right now, the my submission for the applicants on their dashboard is
loaded after 1s of page load which is too much delay.

This PR reduces the delay so it's the loading via htmx is triggered as
soon as the page is finished loading. Also, the PR reduces the loading
preview/skeleton to 3 so there is must of jumpiness before/after the
submissions are loading, this is based on the fact that on an average
there we only few submission by a single applicant.