Fixes to application previews and draft refactoring (#3795)
Fixes #3788 and a few other issues. The main problems being seen around
the preview were when it came to the edit view. This was because the use
of the `Draft` status was being relied on for the saving of content
before previews, when realistically a new application revision had the
same effect.

Previously, Admins/Staff could submit new applications, but could not
submit existing applications. This is why when an Admin had attempted to
submit a preview or submit an edited application, they could not do so.
This functionality has been moved from `ApplicantSubmissionEditView` to
`BaseSubmissionEditView` as it made sense to me that Admins would be
able to access the same edit workflow & transitions as Applicants. I
believe most OTF staff avoid editing incoming applications though.