Add defaults for some messages
This is not all of them, but a good representation for how to use the
variables.  We may want to build more out if the extension gets used
more.

For a large number of the messages, there's non trivial business logic
in either the original messages or the generation of them that needs to
be moved out or handled in another way.

Issue ots/clients/ardc/hypha-tracker#218:  Permit outgoing
messages/notifications to be configured in Wagtail