Make migration resistent to model changes
This migration loads up Users to update their groups.  However, if the
User class has been changed in some way, and the migration runs from a
clean database, it will fail with the error that you cannot select the
new field from the database (as it hasn't been created!)

This change makes it so that only the groups will be loaded, which is
the only part of the User model that is needed for the migration.