Email notifications on new comments for partners (#3913)
Fixes #3871. Allows partners that are assigned to an application to also
get email notifications when a new comment is made that has applicable
visibility. There was some refactoring that took place to allow for
email messages to be customized on a per-recipient basis rather than
having one be generated for all recipients.