Add application submission id everywhere, with an optional prefix (#3930)
Fixes #3830

This PR allow adding an option prefix to the submission ID's that are
auto generated.

The prefix can be added by going to round or lab settings page. 

This PR also ensure the prefix or application ID displayed along with
title almost everywhere for each reference. Since the title of
application is not unique, this helps add uniqueness to it.

The autolinking of submission ids in communication has be updated to
account for application id prefix