Fix #4020: Remove None reviewers from validation
This was caused by an issue where the "---" (None) reviewers were
causing the validation to fail, meaning that you had to fill out n-1
reviewers in the submission page in order for the save to go through.

The solution was to only check for duplicates with reviewers that were
selected.