Sanitize user `full_name` without roadblocks
Rather than impeding the flow of users setting full name or requiring
a new variable to be set by administrators, clean up users' given full
names. This means removing all HTML tags and removing colons and
slashes. It may be unfriendly to people such as "David-Wynn: Miller"
but this is the cost of nerfing spam URLs in full names.

Issue #4060