Make background color consistent across the app (#4059)
The applicant dashboard (and various other views) used tailwind's
`bg-light-grey` which seems inconsistent with the overall app and caused
the menu to have a different standout color.

I believe the grey background does add a nice contrast to things so I
made that the system default. Open to making it all white, though.