Improve cloned forms naming (#4114)
Fixes #3546

Improves the naming of cloned forms by changing the format
to `<original form name> (Copied on <%Y-%m-%d %H:%M:%S.%f>)`, ie.
cloning `Test Form` would result in a duplicated form with the name
`Test Form (Copied on 2024-08-27 17:34:32.004)`. If a form name with a
timestamp already exists on a form it will be subbed via RegEx matching.