Invoice status updating improvements (#4083)
Fixes #3208. Implemented an improvement for the picklist where radio
buttons will be shown when there are less than 4 options. Otherwise the
prompt will be like it was before.

This PR also implements HTMX for the invoice updating prompt, allowing
invoices in table to be updated w/o refreshing the pages

Co-authored-by: Fredrik Jonsson <frjo@xdeb.org>