Add the ability to translate submissions (#4134)
Pertaining to discussions had in #4037. This integrates the
[argostranslate](https://github.com/argosopentech/argos-translate)
library to allow for translation of applications. The current
functionality allows staff users to translate the written answers of an
application.

This feature utilizes htmx to translate the application without
requiring a reload of the page, while also updating the URL so a user
can have a sharable link (using the query parameters `fl` [from
language] & `tl` [to language]) when collaborating with others on the
application.

This feature is disabled by default & can be enabled via the
`SUBMISSION_TRANSLATIONS_ENABLED` setting

## Demos

*Translating an application*


https://github.com/user-attachments/assets/e454a6b7-d449-4caf-8d2d-cb1ad7196546

*Clearing a translation via the form*


https://github.com/user-attachments/assets/5c46ea36-7860-49dc-8758-965a06036fdc