Add a reporting table for projects (#4051)
Fixes #4050

What a small line for such a large set of changes. Those supporting
changes include:

* Creating a view for the new Reporting Table
* Creating a current_report() method in report_config that is slightly
differnt than current_report_due (see inline doc)
* Adding a ProjectQuerySet method that does some intense subquerying to
be able to express the current_report in sql so that the status can be
filtered against


## Note from the cherry-picker/integrator

I see that the summary tables for reports was removed from
`apply/projects` and that the current dropdown in the main navigation
header points to `apply/project/reports`. This new table appears under
`apply/project/reporting` and is linked to in the original fork in which
this change occurred. As for integration here, the question is: do we
want to replace the original table or augment off to the side somehow?
Where should the link to this new table live?

Co-authored-by: Frank Duncan <frankduncan@opentechstrategies.com>
Co-authored-by: Fredrik Jonsson <frjo@xdeb.org>