Fix submission with "invite to proposal" status stuck in draft if language is not english (#4220)
The current implementation tries to find the transition based on the
`name` property of the transitions available. The name of the transition
is a translated string so while submitting a proposal the key "submit"
from the `request.POST` could not be found in the available transition.

This PR updates the logic to fetch the transition target, by searching
through the newly added marker on when to auto transition if the
application is submitted.

Fixes #4211 

See
https://github.com/HyphaApp/hypha/pull/4220/commits/e6656d62f6db57426bb15f298127c106ea4f6397
for patch.