Separate SOW & PAF form editing (#4261)
Fixes #4204. Moves the SOW out into it's own form when editing. This
also removes the `user_has_updated_details` attribute in favor of using
a property for both the project form & SOW that checks the field_data on
the respective form. This allows for tracking of the project form & SOW
independently.

Also a few small aesthetic changes bundled in, like margin additions &
hiding of submission attachments sidebar when there's none.

---------

Co-authored-by: Fredrik Jonsson <frjo@xdeb.org>