Added ability to set initial status when creating a project (#4390)
Closes #4388. Adds a new field to the "Create Project" modal that allows
for setting the initial status of the project. The default for this
field can be set with the `PROJECTS_DEFAULT_STATUS` environment setting,
which would also set the initial status when `PROJECTS_AUTO_CREATE` is
enabled. This settings needs to match one of the string literals here:

https://github.com/HyphaApp/hypha/blob/cc174b61b6af7dfd9403cdb2fe3a13a9364e37f7/hypha/apply/projects/models/project.py#L75-L80

In the modal the statuses of `INTERNAL_APPROVAL` and `COMPLETE` were
removed as choices as I didn't see them being useful statuses to be
manually set (internal approval would only be approving the items
created in the `DRAFT` step)

Also updated the activity message to reflect what status the project was
created with