Merge application and project communication tabs (#4349)
Fixes #4249. Migrates all project comments to the associated
application, and loads the applications comments via htmx (merging in
project activities). This was the best solution I could come up with
until we potentially take a step a coupling the two closer (ie. sharing
an ID (#3944), a URL space, and/or a detail view).

I had tried another version of this that did the same migration but
loaded the project view via htmx from the application detail view, which
made everything feel more consistent, but I had concerns about the user
needed to see the status/stages bar when making a comment which is where
the UI got hairy.