Update public facing id of project to be the id of the submission (#3944)
Update the project urls to use it's submission id instead of the
project's db id

The reason for the update is that there are two main entities in the
system, one submission and other project.
The ids of the submission is linkable in the communiction and also used
for referencing at different places.

Considering a project is an extension of application submission, if we
have the project publically available id same as that of it's submission
it always to link and reference the project by the same id.


## Test Steps
<!-- 
If step does not require manual testing, skip/remove this section.

Give a brief overview of the steps required for a user/dev to test this
contribution. Important things to include:
 - Required user roles for where necessary (ie. "As a Staff Admin...")
- Clear & validatable expected results (ie. "Confirm the submit button
is now not clickable")
- Language that can be understood by non-technical testers if being
tested by users
-->

- [ ] General regression test and make sure all the urls in the project
section resolve and link to the right pages