Block multiple form submits on application form. (#4452)
Fixes #4348

Before this js snippet it was possible on a slow network to post a
application form multiple times by rapidly clicking the submit button
multiple times.

Would be neat to use aplinejs for this everywhere but it interfere with
the current "unlockApplicationForm" feature that we use to block some
spam.