Small improvements for displaying closed funds (#4456)
A couple simple additions:
- Added a new message (as a customizable system setting) that will
display when there are no open funds
- Added call to action to closed round message along with a fix for
displaying the verbose name
- Small misc. fixes/additions (`Mobil` -> `Mobile` typo in system
settings, docstrings in `util_tags.py`

## Test Steps
<!-- 
If step does not require manual testing, skip/remove this section.

Give a brief overview of the steps required for a user/dev to test this
contribution. Important things to include:
 - Required user roles for where necessary (ie. "As a Staff Admin...")
- Clear & validatable expected results (ie. "Confirm the submit button
is now not clickable")
- Language that can be understood by non-technical testers if being
tested by users
-->

- [ ] Ensure that the set system settings message displays on the home
page when no funds are live
- [ ] Ensure there is a link to the homepage in the message displayed
when a closed round URL is accessed

---------

Co-authored-by: Saurabh Kumar <theskumar@users.noreply.github.com>