Modernize the task-list UI
### Notable changes:

- Show tasks in chronological order
- Add relative timestamp against each tasks
- Review view button to save some spaces and make the row clickable
- Update the tasks copy to include the fund name and application id

### Worth considering (up for discussion):

- should be update and make use of `[fund_name] #{application_id}` as
generic indentier for applications instead of the current "title" only.
- When it comes to selecting a main fund name, should use priorities
name of the "round" or name of the "fund" from which the round is
derived.

## Internal changes

Included `fund_name` and `application_id` on both the submission and
project instance to easily access them, while abstracting the complexity
around fallback on submission.id when the submission.public_id is
missing