Refactor "Reports" into its own app (#4468)
- The app is named "project_reports," and currently, the table names are
hardcoded.
- It is currently located under "hypha.apply.projects.reports," but in
the future, I plan to move the entire "projects" app/module as
"hypha.projects."
- The database tables will stay unchanged; the migration state will show
that the models have been relocated to a different app.
- All URLs will stay the same.

This is the initial step among other refactors/fixes for the reporting
module.

Additionally, once this is approved, I will proceed to refactor the
invoicing module.